Update deps, fix tests
