+++ orchestration/hca_orchestration/tests/solids/test_create_snapshot.py
@@ -1,11 +1,16 @@
 import unittest
+from dagster_utils.resources.slack import console_slack_client
 from unittest.mock import patch
 
-from dagster import execute_solid
+from dagster import execute_solid, ModeDefinition, ResourceDefinition
 
+from hca_orchestration.config import preconfigure_resource_for_mode
 from hca_manage.common import JobId
 from hca_orchestration.pipelines.cut_snapshot import test_mode
 from hca_orchestration.solids.create_snapshot import make_snapshot_public, submit_snapshot_job
+from hca_orchestration.resources.config.data_repo import hca_manage_config, snapshot_creation_config
+
+from hca_orchestration.resources.config.dagit import dagit_config
 
 
 class CreateSnapshotSolidsTestCase(unittest.TestCase):
@@ -32,7 +37,7 @@ class CreateSnapshotSolidsTestCase(unittest.TestCase):
             submit_snap.assert_called_once_with('namityname', False)
 
     def test_make_snapshot_public_hits_correct_sam_path(self):
-        with patch('dagster_utils.resources.sam.NoopSamClient.make_snapshot_public') as mock_make_public:
+        with patch('dagster_utils.resources.sam.NoopSamClient.set_public_flag') as mock_set_public:
             result = execute_solid(
                 make_snapshot_public,
                 run_config={},
@@ -41,4 +46,4 @@ class CreateSnapshotSolidsTestCase(unittest.TestCase):
                 },
                 mode_def=test_mode)
             self.assertTrue(result.success)
-            mock_make_public.assert_called_once_with('steve')
+            mock_set_public.assert_called_once_with('steve', True)
