+++ orchestration/pyproject.toml
@@ -20,7 +20,7 @@ python-dateutil = "^2.8.1"
 typing-extensions = "^3.7.4"
 pyyaml = "^5.3"
 dagster-gcp = "^0.12.10"
-broad-dagster-utils = "0.6.5"
+broad-dagster-utils = "0.6.6"
 hca-import-validation = "^0.0.1"
 more-itertools = "^8.8.0"
 graphql-ws = "<0.4.0"
